+++ lib/flutter_native_getx_controller.dart
@@ -120,6 +120,10 @@ class FlutterNativeGetxController extends SuperController {
   }
 
   void _updateEventTypeFinished() {
+    controllerTimeout?.cancel();
+    isShowLoading = false;
+    isVisibleButtonPlay = true;
+    isShowController = true;
     iconControlPlayer = const Icon(
       Icons.replay,
       color: Colors.white,