+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 1.0.1
 
 * Fix the issue can not use the plugin on a project that has version Kotlin lower than 1.6.10
+
+## 1.0.2
+
+* Change Minimum plugin require on iOS from iOS 10 to iOS 9.
+* Fix the issue playback state on iOS does not work properly.